Restructuring the Rust module and adding pre-commit hooks for Rust
